Extract global ContextState object to doctest/parts/context_state.h
